requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15,<3.17
requests>=2.31,<3
pytz>=2024.1
